Integrate http request completion with async call
