+++ Sources/WalletConnectRelay/HTTP/HTTPClient.swift
@@ -1,5 +1,10 @@
 import Foundation
 
+struct Endpoint {
+    let path: String
+    let queryParameters: [URLQueryItem]
+}
+
 final class HTTPClient {
 
     let host: String
@@ -11,14 +16,27 @@ final class HTTPClient {
         self.session = session
     }
 
-    func request<T: Decodable>(_ endpoint: Endpoint, completion: @escaping (HTTPResponse<T>) -> Void) {
+    func request<T: Decodable>(_ type: T.Type, at endpoint: Endpoint) async throws -> T {
+        return try await withCheckedThrowingContinuation { continuation in
+            request(T.self, at: endpoint) { response in
+                do {
+                    let value = try response.result.get()
+                    continuation.resume(returning: value)
+                } catch {
+                    continuation.resume(throwing: error)
+                }
+            }
+        }
+    }
+
+    func request<T: Decodable>(_ type: T.Type, at endpoint: Endpoint, completion: @escaping (HTTPResponse<T>) -> Void) {
         let request = makeRequest(for: endpoint)
         session.dataTask(with: request) { data, response, error in
             completion(HTTPResponse(request: request, data: data, response: response, error: error))
         }.resume()
     }
 
-    func makeRequest(for endpoint: Endpoint) -> URLRequest {
+    private func makeRequest(for endpoint: Endpoint) -> URLRequest {
         var components = URLComponents()
         components.scheme = "https"
         components.host = host
@@ -30,61 +48,5 @@ final class HTTPClient {
         var request = URLRequest(url: url)
         request.httpMethod = "GET"
         return request
-    }
-}
-
-enum HTTPError: Error {
-    case dataTaskError(Error)
-    case noResponse
-    case badStatusCode(Int)
-    case responseDataNil
-    case jsonDecodeFailed(Error, Data)
-}
-
-struct HTTPResponse<T: Decodable> {
-
-    let request: URLRequest?
-    let data: Data?
-    let urlResponse: HTTPURLResponse?
-    let result: Result<T, Error>
-
-    init(request: URLRequest? = nil, data: Data? = nil, response: URLResponse? = nil, error: Error? = nil) {
-        self.data = data
-        self.request = request
-        self.urlResponse = response as? HTTPURLResponse
-        self.result = Self.validate(data, response, error).flatMap { data -> Result<T, Error> in
-            if let rawData = data as? T {
-                return .success(rawData)
-            }
-            return Self.decode(data)
-        }
-    }
-}
-
-extension HTTPResponse {
-
-    private static func validate(_ data: Data?, _ urlResponse: URLResponse?, _ error: Error?) -> Result<Data, Error> {
-        if let error = error {
-            return .failure(HTTPError.dataTaskError(error))
-        }
-        guard let httpResponse = urlResponse as? HTTPURLResponse else {
-            return .failure(HTTPError.noResponse)
-        }
-        guard (200..<300) ~= httpResponse.statusCode else {
-            return .failure(HTTPError.badStatusCode(httpResponse.statusCode))
-        }
-        guard let validData = data else {
-            return .failure(HTTPError.responseDataNil)
-        }
-        return .success(validData)
-    }
-
-    private static func decode<T: Decodable>(_ data: Data) -> Result<T, Error> {
-        do {
-            let decoded = try JSONDecoder().decode(T.self, from: data)
-            return .success(decoded)
-        } catch let jsonError {
-            return .failure(HTTPError.jsonDecodeFailed(jsonError, data))
-        }
     }
 }
